+++ util/util.py
@@ -395,3 +395,15 @@ def check_for_kb_interrupt():
         raise current_exc
     return
 
+def add_packages(list1, list2):
+    # This function dedups the package names which I think could be
+    # functionally not desirable however rather than changing the behavior
+    # instead we will do the dedup in a stable manner that prevents
+    # package re-ordering
+    pkgs = {re.sub('==.+', '', pkg): pkg for pkg in list1 + list2}
+    merged = []
+    for k in list1 + list2:
+        v = pkgs.pop(re.sub('==.+', '', k), None)
+        if v is not None:
+            merged.append(v)
+    return merged
